Update venue list for returning users.
This could be taken out of the request cycle and deferred using something like Celery, but given that it's quicker for returning users (because we only look at new checkins, not all checkins) this won't do much harm. And I went to a nice lunch place for the first time today, so it should be in here.